create class teacher.rb
