--- Person.rb
@@ -1,24 +1,0 @@
-class Person
-	attr_accessor :name, :age, :parent_permission
-	attr_reader :id
-
-	def initialize(age:, name:"Unknown", parent_permission:true)
-		@name = name
-		@age = age
-		@parent_permission = parent_permission
-
-	end
-
-	def can_use_services?
-			self.is_of_age? || @parent_permission
-	end
-	
-	private
-	
-	def is_of_age?
-		@age >= 18
-	end
-end
-
-
-
